backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
gunicorn==21.2.0
chromadb==0.4.15
sentence-transformers==2.2.2
transformers==4.35.2
torch==2.1.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
requests==2.31.0
python-dotenv==1.0.0
huggingface-hub==0.16.4
accelerate==0.24.1
